LayoutTests/fast/forms/checkValidity-002-expected.txt

This test checks if checkValidity() returns correctly a false (meaning error) result on invalid elements, and returns a true result on a blank but valid elements. Blank but non-placeholder label options are valid.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".



PASS checkValidityFor("input-empty") is false
PASS checkValidityFor("input-pattern-mismatch") is false
PASS checkValidityFor("textarea") is false
PASS checkValidityFor("select-no-explicit-value") is true
PASS checkValidityFor("select-placeholder") is false
PASS checkValidityFor("select-non-placeholder") is true
PASS successfullyParsed is true

TEST COMPLETE
